qsort: Add configure option to use median pivot
The configure script now has a --enable-media-pivot option to
use the median of the first, middle or last value in the
array as the pivot, as suggested by Sedgewick.

Also added a simple random seed for generating the list to sort,
so we don't get the same list of numbers on each run.

* configure.ac: Add --enable-median-pivot option
* qsort.c
  * setpivot(): New function to choose pivot. Support configurable
    MEDIAN_PIVOT method
  * quicksort(): Call setpivot() to establish the sort pivot
  * main(): Simple srand() seed to get different sort lists